nim_app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
langchain-nvidia-ai-endpoints>=0.1.0
pydantic>=2.0.0,<3.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0